Adding static imports.

+++ src/main/java/edu/emory/mathcs/csparsej/tdcomplex/DZcs_cumsum.java
@@ -22,7 +22,7 @@
  *
  */
 
-package edu.emory.mathcs.csparsej.tdcomplex;
+package edu.emory.mathcs.csparsej.tdcomplex ;
 
 /**
  * Cumulative sum.
@@ -33,30 +33,32 @@ package edu.emory.mathcs.csparsej.tdcomplex;
  */
 public class DZcs_cumsum {
 
-    /**
-     * p [0..n] = cumulative sum of c [0..n-1], and then copy p [0..n-1] into c
-     *
-     * @param p
-     *            size n+1, cumulative sum of c
-     * @param c
-     *            size n, overwritten with p [0..n-1] on output
-     * @param n
-     *            length of c
-     * @return sum (c), null on error
-     */
-    public static int cs_cumsum(int[] p, int[] c, int n) {
-        int i, nz = 0;
-        double nz2 = 0;
-        if (p == null || c == null)
-            return (-1); /* check inputs */
-        for (i = 0; i < n; i++) {
-            p[i] = nz;
-            nz += c[i];
-            nz2 += c[i]; /* also in double to avoid int overflow */
-            c[i] = p[i]; /* also copy p[0..n-1] back into c[0..n-1]*/
-        }
-        p[n] = nz;
-        return (int) nz2; /* return sum (c [0..n-1]) */
-    }
+	/**
+	 * p [0..n] = cumulative sum of c [0..n-1], and then copy p [0..n-1] into c
+	 *
+	 * @param p
+	 *            size n+1, cumulative sum of c
+	 * @param c
+	 *            size n, overwritten with p [0..n-1] on output
+	 * @param n
+	 *            length of c
+	 * @return sum (c), null on error
+	 */
+	public static double cs_cumsum(int [] p, int [] c, int n)
+	{
+		int i, nz = 0 ;
+		double nz2 = 0 ;
+		if (p == null || c == null)
+			return (-1) ;		/* check inputs */
+		for (i = 0 ; i < n ; i++)
+		{
+			p [i] = nz ;
+			nz += c [i] ;
+			nz2 += c [i] ;		/* also in double to avoid int overflow */
+			c [i] = p [i] ;		/* also copy p[0..n-1] back into c[0..n-1]*/
+		}
+		p [n] = nz ;
+		return (nz2) ;			/* return sum (c [0..n-1]) */
+	}
 
 }
